+++ regression/reference/scope/wrapcls2Enum.h
@@ -1,0 +1,49 @@
+// wrapcls2Enum.h
+// This is generated code, do not edit
+// #######################################################################
+// Copyright (c) 2017-2019, Lawrence Livermore National Security, LLC.
+//
+// Produced at the Lawrence Livermore National Laboratory
+//
+// LLNL-CODE-738041.
+//
+// All rights reserved.
+//
+// This file is part of Shroud.
+//
+// For details about use and distribution, please read LICENSE.
+//
+// #######################################################################
+/**
+ * \file wrapcls2Enum.h
+ * \brief Shroud generated wrapper for cls2Enum class
+ */
+// For C users and C++ implementation
+
+#ifndef WRAPCLS2ENUM_H
+#define WRAPCLS2ENUM_H
+
+#include "typesscope.h"
+
+// splicer begin class.cls2Enum.CXX_declarations
+// splicer end class.cls2Enum.CXX_declarations
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+//  cls2Enum::Color
+enum SCO_cls2Enum_Color {
+    SCO_cls2Enum_Color_RED = 50,
+    SCO_cls2Enum_Color_BLUE,
+    SCO_cls2Enum_Color_WHITE
+};
+
+// splicer begin class.cls2Enum.C_declarations
+// splicer end class.cls2Enum.C_declarations
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif  // WRAPCLS2ENUM_H